Resolve compile errors in tests

+++ test/chow/zidane/ojs/leetcode/FirstMissingPositiveTest.java
@@ -11,7 +11,7 @@ class FirstMissingPositiveTest {
     @ParameterizedTest
     @CsvSource( {"'0, 1', 2", "'0', 1", "'1', 2", "'2, 4, 5, 99', 1", "'23, 41, 32, 0, 1', 2", "'3, 4, -1, 1', 2", "'23, 24, 1, 12', 2"})
     void test(@ConvertWith(StringToIntArrayConverter.class) final int[] nums, final int out) {
-        assertEquals(out, FirstMissingPositive.MaximumRange.find(nums));
-        assertEquals(out, FirstMissingPositive.Swap.find(nums));
+        assertEquals(out, FirstMissingPositive.MaximumRange.find(nums).intValue());
+        assertEquals(out, FirstMissingPositive.Swap.find(nums).intValue());
     }
 }
