+++ test/chow/zidane/ojs/leetcode/MissingNumberTest.java
@@ -11,7 +11,7 @@ class MissingNumberTest {
     @ParameterizedTest
     @CsvSource( {"'0',1", "'1',0", "'0,1',2", "'2,3,4,0',1", "'3,4,2,0',1", "'5,4,1,2,0',3"})
     void test(@ConvertWith(StringToIntArrayConverter.class) final int[] in, final int out) {
-        assertEquals(out, MissingNumber.SumSubtraction.find(in));
-        assertEquals(out, MissingNumber.ExclusiveOr.find(in));
+        assertEquals(out, MissingNumber.SumSubtraction.find(in).intValue());
+        assertEquals(out, MissingNumber.ExclusiveOr.find(in).intValue());
     }
 }
