+++ test/chow/zidane/ojs/leetcode/SingleNumberTest.java
@@ -11,6 +11,6 @@ class SingleNumberTest {
     @ParameterizedTest
     @CsvSource( {"'3,2,2,4,4',3", "0,0,-1,-1,2,2,-3',-3", "'2,2,4',4", "'3,2,2,4,4',3", "'1,0,1',0"})
     void test(@ConvertWith(StringToIntArrayConverter.class) final int[] nums, final int out) {
-        assertEquals(out, SingleNumber.ExclusiveOr.find(nums));
+        assertEquals(out, SingleNumber.ExclusiveOr.find(nums).intValue());
     }
 }
