+++ test/chow/zidane/ojs/leetcode/ThirdMaximumNumberTest.java
@@ -11,6 +11,6 @@ class ThirdMaximumNumberTest {
     @ParameterizedTest
     @CsvSource( {"'1,2',2", "'1,2,3',1", "'1,1,1,1,1',1", "'1,2,3,4,5,4',3", "'1,2,3,4,5',3"})
     void test(@ConvertWith(StringToIntArrayConverter.class) final int[] nums, final int out) {
-        assertEquals(out, ThirdMaximumNumber.StoreTop3Max.find(nums));
+        assertEquals(out, ThirdMaximumNumber.StoreTop3Max.find(nums).intValue());
     }
 }
